[ETCM-1096] Improve SignedTransactionSpec for valid pointSigns
